requirements.txt.txt
Flask
google-generativeai
Pillow
gunicorn